vqa/requirements.txt
auto_mix_prep==0.2.0
google-genai
numpy==2.3.5
ollama==0.6.1
Pillow==12.0.0
protobuf==6.33.1
pydantic==2.12.5
PyYAML==6.0.3
scipy==1.16.3
tqdm==4.67.1
